+++ packages/sane/lib/sane.dart
@@ -1,8 +1,8 @@
 library;
 
 export 'src/exceptions.dart';
-export 'src/sane_dev.dart';
-export 'src/sane_native.dart';
-export 'src/sane_sync.dart';
+export 'src/impl/sane_dev.dart';
+export 'src/impl/sane_native.dart';
+export 'src/impl/sane_sync.dart';
 export 'src/structures.dart';
 export 'src/utils.dart';
