+++ packages/sane/lib/src/isolate_messages/interface.dart
@@ -1,4 +1,4 @@
-import 'package:sane/src/sane_sync.dart';
+import 'package:sane/src/impl/sane_sync.dart';
 
 abstract interface class IsolateMessage<T extends IsolateResponse> {
   Future<T> handle(Sane sane);
