+++ packages/sane/lib/src/isolate_messages/read.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
+import 'package:sane/src/impl/sane_sync.dart';
 import 'package:sane/src/isolate_messages/interface.dart';
-import 'package:sane/src/sane_sync.dart';
 import 'package:sane/src/structures.dart';
 
 class ReadMessage implements IsolateMessage<ReadResponse> {
