+++ packages/sane/lib/src/isolate_messages/set_io_mode.dart
@@ -1,5 +1,5 @@
+import 'package:sane/src/impl/sane_sync.dart';
 import 'package:sane/src/isolate_messages/interface.dart';
-import 'package:sane/src/sane_sync.dart';
 import 'package:sane/src/structures.dart';
 
 class SetIOModeMessage implements IsolateMessage {
