+++ packages/sane/test/sane_test.dart
@@ -1,9 +1,8 @@
-import 'package:sane/src/sane_sync.dart';
+import 'package:sane/src/impl/sane_sync.dart';
 import 'package:test/test.dart';
 
 void main() {
   test('Sane init test', () {
-    final sane = Sane();
-    expect(sane.init, returnsNormally);
+    expect(Sane.new, returnsNormally);
   });
 }
